keep default on double precision for now

+++ src/Wflow.jl
@@ -18,7 +18,7 @@ using CFTime
 @metadata get_units "mm"
 
 const BMI = BasicModelInterface
-const Float = Float32
+const Float = Float64
 
 mutable struct Clock{T}
     time::T
